vcms/requirements.txt
django>=1.9
Pillow
sorl-thumbnail
MySQL-python
django-mptt
django_rq
rq-scheduler
markdown
django-filter
django-crispy-forms
#django-supervisor
python-dateutil
mandrill
wget
django-suit
django-suit-rq
django-reversion
django-suit-redactor
django-model-utils
django-taggit
django_select2
django-location-field
shortuuid
django-autoslug
django-redis
django-redis-cache
jsonfield
djangorestframework
django-haystack
libsass
elasticsearch>=1.0.0,<2.0.0
